Changes to support building with github

+++ kafkaBuild.bat
@@ -1,16 +1,14 @@
 :PREP
 
-	for /f "tokens=1,2 delims=_" %%a in ("%JOB_NAME%") do set tsk=%%a&set vers=%%b
+	for /f "tokens=2,3 delims=/" %%a in ("%1") do set prod=%%a&set branch=%%b
 
-	set TASK=%tsk%
-	set BRANCH=%vers%
-
+	echo Building %prod% from %branch%
 	GOTO VS
 
 
 :VS
-
-	IF "%BRANCH%" == "Trunk" (
+	GOTO USE_VS17
+	IF "%branch%" == "main" (
 		GOTO USE_VS17
 	)
 	IF EXIST "c:\Program Files (x86)\Microsoft Visual Studio 8" GOTO USE_VS8
@@ -41,7 +39,7 @@
 :BUILD
 
 	set WORKSPACE=%CD%
-	rmdir /q /s %TEMP%\kafka
+	IF EXIST "%TEMP%\kafka" rmdir /q /s %TEMP%\kafka
 
 	dotnet new classlib  --name kafka -o %TEMP%\kafka --force
 	dotnet add %TEMP%\kafka package librdkafka.redist --version 2.5.0
@@ -49,7 +47,7 @@
 	
 	xcopy %USERPROFILE%\.nuget\packages\librdkafka.redist\2.5.0 .\packages\librdkafka.redist.2.5.0\	/S /E /Y
 
-	rmdir /q /s %TEMP%\kafka
+	IF EXIST "%TEMP%\kafka" rmdir /q /s %TEMP%\kafka
 
 	for /f "skip=1" %%d in ('wmic os get localdatetime') do if not defined mydate set mydate=%%d
 	set BDATE=%mydate:~0,8%-%mydate:~8,4%
@@ -80,5 +78,5 @@
 	set /a RESULT=1
 
 :End
-	rmdir /q /s .\packages
+	IF EXIST ".\packages" rmdir /q /s .\packages
 	EXIT /B %RESULT%
